chore: Setup express and dev tools
This commit adds dependencies for express and type definitions for Node
and Express. The express library will be used to setup the server. This
is deffininately overkill for only one route, but this project is about
figuring out the setup for larger projects built with Typescript and
Express.

Eslint, Prettier, and necessary configuration is setup to use Eslint
for code style checking and Prettier to auto-format. VSCode or a
compatible IDE should be configured to auto fix linting errors on save.

Jest is setup for running tests and coverage reporting. Superagent is
used to provide a clean API for testing Express routes.

- Adds type defs for Node, Express, Jest, and Supertest.
- Typescript build output is stored in dist/